Moving filebackupstorage into its own package.
